refactor: split logger.rs into modular components
- Split the 2100+ line logger.rs file into dedicated modules
- Organize code into environment, error, init, and path_resolver modules
- Fixed doctest issues with SecretString API key initialization
- All tests passing with improved organization and maintenance
