+++ benches/simple_benchmark.rs
@@ -4,6 +4,7 @@
 use bytes::Bytes;
 use hyper::{HeaderMap, Method, Uri};
 use reqwest::StatusCode;
+use secrecy::SecretString;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 use switchboard::config::Config;
@@ -73,7 +74,7 @@ fn run_benchmark(mode: LogMode, iterations: usize, body: &Bytes) -> Duration {
         LogMode::StdoutOnly => {
             let config = Arc::new(Config {
                 port: "8080".to_string(),
-                anthropic_api_key: "test".to_string(),
+                anthropic_api_key: SecretString::new("test".to_string().into()),
                 anthropic_target_url: "https://example.com".to_string(),
                 openai_api_key: None,
                 openai_api_base_url: "https://api.openai.com".to_string(),
@@ -98,7 +99,7 @@ fn run_benchmark(mode: LogMode, iterations: usize, body: &Bytes) -> Duration {
         LogMode::FileOnly => {
             let config = Arc::new(Config {
                 port: "8080".to_string(),
-                anthropic_api_key: "test".to_string(),
+                anthropic_api_key: SecretString::new("test".to_string().into()),
                 anthropic_target_url: "https://example.com".to_string(),
                 openai_api_key: None,
                 openai_api_base_url: "https://api.openai.com".to_string(),
@@ -126,7 +127,7 @@ fn run_benchmark(mode: LogMode, iterations: usize, body: &Bytes) -> Duration {
         LogMode::DualOutput => {
             let config = Arc::new(Config {
                 port: "8080".to_string(),
-                anthropic_api_key: "test".to_string(),
+                anthropic_api_key: SecretString::new("test".to_string().into()),
                 anthropic_target_url: "https://example.com".to_string(),
                 openai_api_key: None,
                 openai_api_base_url: "https://api.openai.com".to_string(),
@@ -222,7 +223,7 @@ fn test_nonblocking_io() {
     // Set up file logging
     let config = Arc::new(Config {
         port: "8080".to_string(),
-        anthropic_api_key: "test".to_string(),
+        anthropic_api_key: SecretString::new("test".to_string().into()),
         anthropic_target_url: "https://example.com".to_string(),
         openai_api_key: None,
         openai_api_base_url: "https://api.openai.com".to_string(),
