+++ tests/common/mod.rs
@@ -2,6 +2,7 @@
 
 use axum::Router;
 use reqwest::Client;
+use secrecy::SecretString;
 use std::fs;
 use std::io;
 use std::io::{BufRead, BufReader};
@@ -69,17 +70,17 @@ pub async fn setup_test_environment() -> TestSetup {
     // Use dummy values for fields that are suitable for testing
     let config = Config {
         port: "0".to_string(), // Use 0 to let OS assign a random port if needed
-        anthropic_api_key: "test-api-key".to_string(), // Dummy API key for testing
+        anthropic_api_key: SecretString::new("test-api-key".to_string().into()), // Dummy API key for testing
         anthropic_target_url: mock_server.uri(), // Point to the mock server
-        openai_api_key: Some("test-openai-api-key".to_string()), // Dummy OpenAI API key
+        openai_api_key: Some(SecretString::new("test-openai-api-key".to_string().into())), // Dummy OpenAI API key
         openai_api_base_url: "https://api.openai.com".to_string(), // Default OpenAI URL
-        openai_enabled: false, // Disabled by default
+        openai_enabled: false,                                     // Disabled by default
         log_stdout_level: "debug".to_string(), // Use debug level for more verbose test logs
-        log_format: "pretty".to_string(), // Use pretty format for readability in tests
-        log_bodies: true,      // Enable body logging for verbose testing
+        log_format: "pretty".to_string(),      // Use pretty format for readability in tests
+        log_bodies: true,                      // Enable body logging for verbose testing
         log_file_path: "./test-switchboard.log".to_string(), // Test-specific log file
-        log_file_level: "trace".to_string(), // Most verbose for file logs in tests
-        log_max_body_size: 20480, // Default size for tests
+        log_file_level: "trace".to_string(),   // Most verbose for file logs in tests
+        log_max_body_size: 20480,              // Default size for tests
         log_directory_mode: switchboard::config::LogDirectoryMode::Default, // Use automatic detection for tests
         log_max_age_days: None,
     };
@@ -130,9 +131,9 @@ pub fn setup_test_logging(test_name: &str) -> TestLoggingSetup {
     // Create a test-specific configuration
     let config = Config {
         port: "0".to_string(),
-        anthropic_api_key: "test-api-key".to_string(),
+        anthropic_api_key: SecretString::new("test-api-key".to_string().into()),
         anthropic_target_url: "https://example.com".to_string(),
-        openai_api_key: Some("test-openai-api-key".to_string()),
+        openai_api_key: Some(SecretString::new("test-openai-api-key".to_string().into())),
         openai_api_base_url: "https://api.openai.com".to_string(),
         openai_enabled: false,
         log_stdout_level: "debug".to_string(),
@@ -182,9 +183,9 @@ pub fn generate_test_log_path(test_name: &str) -> PathBuf {
     // Create a dummy config for the resolver
     let config = Config {
         port: "0".to_string(),
-        anthropic_api_key: "test-api-key".to_string(),
+        anthropic_api_key: SecretString::new("test-api-key".to_string().into()),
         anthropic_target_url: "https://example.com".to_string(),
-        openai_api_key: Some("test-openai-api-key".to_string()),
+        openai_api_key: Some(SecretString::new("test-openai-api-key".to_string().into())),
         openai_api_base_url: "https://api.openai.com".to_string(),
         openai_enabled: false,
         log_stdout_level: "debug".to_string(),
@@ -238,9 +239,9 @@ pub fn verify_log_directory() -> bool {
     // Create a dummy config with LogDirectoryMode::Default
     let config = Config {
         port: "0".to_string(),
-        anthropic_api_key: "test-key".to_string(),
+        anthropic_api_key: SecretString::new("test-key".to_string().into()),
         anthropic_target_url: "https://example.com".to_string(),
-        openai_api_key: Some("test-openai-api-key".to_string()),
+        openai_api_key: Some(SecretString::new("test-openai-api-key".to_string().into())),
         openai_api_base_url: "https://api.openai.com".to_string(),
         openai_enabled: false,
         log_stdout_level: "info".to_string(),
