+++ tests/directory_creation_test.rs
@@ -1,3 +1,4 @@
+use secrecy::SecretString;
 use std::{env, fs, path::PathBuf};
 use switchboard::{
     config::{Config, LogDirectoryMode, DEFAULT_LOG_DIRECTORY_MODE},
@@ -13,11 +14,11 @@ use std::os::unix::fs::PermissionsExt;
 /// Helper function to create a test config with specific settings
 fn create_test_config(log_file_path: &str, log_directory_mode: LogDirectoryMode) -> Config {
     Config {
-        openai_api_key: Some("test-openai-api-key".to_string()),
+        openai_api_key: Some(SecretString::new("test-openai-api-key".to_string().into())),
         openai_api_base_url: "https://api.openai.com".to_string(),
         openai_enabled: false,
         port: "0".to_string(),
-        anthropic_api_key: "test-api-key".to_string(),
+        anthropic_api_key: SecretString::new("test-api-key".to_string().into()),
         anthropic_target_url: "https://example.com".to_string(),
         log_stdout_level: "info".to_string(),
         log_format: "pretty".to_string(),
